Refactor SSD1306 Driver   * Add #defines for all init cmds   * Create new SSD1306Config.h to decouple high level SSD1306 from hardware layer
