+++ STM32F407G-DISC1.xml
@@ -3,8 +3,8 @@
   <board id="stm32f407g-disc1">
     <name>STM32F407G-DISC1</name>
     <mcuId>stm32f407vgtx</mcuId>
-    <dbgIF>SWD</dbgIF>
     <dbgIF>JTAG</dbgIF>
+    <dbgIF>SWD</dbgIF>
     <dbgDEV>ST-LinkV2-1</dbgDEV>
   </board>
 </targetDefinitions>
