+++ src/main.c
@@ -30,83 +30,6 @@
 // Private Function Declarations
 static void STM32_initAllPeripherals(void);
 
-// Private Data
-uint8_t global_display_buffer[((128*64)/8)] = {
-
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,        // Page 0 (Top Most)
-    0x80,0xC0,0x40,0x60,0x20,0x20,0x20,0x20,0x60,0xC0,0x80,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x80,0x80,0x00,0x00,0x00,0x00,0xF8,0x7C,0x86,
-    0xF6,0x1E,0x04,0x0C,0xF8,0xE0,0xC0,0x30,0x88,0xCC,0x24,0x26,0x06,0x06,0xFC,0xF0,
-    0x00,0x00,0x80,0xC0,0x80,0xC0,0xC0,0x80,0x80,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-
-    0x00,0x00,0x00,0x00,0x00,0x60,0xFC,0x0C,0x7A,0x0F,0x81,0x03,0x02,0x06,0x0C,0x1C,
-    0x07,0x00,0xC0,0x43,0x01,0x01,0x03,0x04,0x00,0x21,0x3F,0x1E,0x18,0x08,0x0C,0x08,
-    0x0C,0xC8,0x18,0x18,0x70,0xE0,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x80,0xC0,0x60,0xE0,0xC0,0x40,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x3C,0xF6,0xC2,0xA3,0x61,0xC1,0xC1,0x81,0x03,0x03,0x06,0x1F,0x3A,0x21,
-    0x81,0x00,0x00,0x00,0x07,0x7F,0xFF,0x1E,0x02,0x00,0x00,0xE0,0x70,0x3E,0x0F,0x06,
-    0x83,0x81,0xC1,0x60,0x60,0x70,0x78,0x39,0x27,0x73,0x20,0x60,0xC0,0xC0,0x00,0x00,
-
-    0x00,0x00,0x00,0x00,0x80,0xC0,0x31,0x1B,0x0E,0x0C,0x1C,0x00,0x18,0x10,0x30,0x30,
-    0x00,0x00,0x83,0xFE,0x3C,0x18,0x18,0x70,0xE0,0x30,0x00,0x00,0x18,0x00,0x00,0x08,
-    0x08,0x00,0x0C,0x4E,0xDB,0x30,0x30,0x60,0xC0,0x80,0x00,0x00,0x00,0xC0,0xA0,0x98,
-    0xA4,0xB4,0xAC,0x08,0x30,0x30,0xC0,0x80,0x00,0x00,0x00,0x00,0x00,0x00,0xD0,0x38,
-    0x0E,0x0E,0x07,0x07,0x00,0x00,0x00,0x0F,0x07,0x02,0x33,0x99,0xEF,0x0B,0x00,0x00,
-    0x00,0x80,0xC0,0x60,0x30,0x19,0x0B,0x0B,0x0E,0x06,0x0E,0x44,0x06,0x06,0x06,0x86,
-    0x84,0x8C,0xEC,0xF8,0x0C,0xE6,0xB7,0x86,0x04,0x0C,0x79,0xF8,0x88,0x0C,0x02,0x86,
-    0x05,0x00,0x00,0x02,0xC2,0xE1,0xE1,0xE1,0xA3,0xB2,0x30,0x10,0x18,0x1F,0x07,0x00,
-
-    0x00,0x00,0x00,0x00,0x0F,0x39,0x70,0x60,0xC0,0xC0,0xC0,0xE0,0x00,0x00,0x80,0x10,
-    0x18,0x0C,0x06,0x07,0x06,0x04,0x7E,0x67,0x07,0x06,0x04,0x0C,0x0C,0x0C,0x48,0x40,
-    0xE0,0x72,0x60,0x62,0x23,0x31,0x30,0x18,0x0F,0x03,0x00,0x00,0x01,0x07,0x01,0x01,
-    0x00,0x00,0x01,0x03,0x06,0x9C,0x60,0x01,0x06,0xDC,0x70,0x00,0x00,0xED,0xBF,0xE0,
-    0xE0,0xF0,0x70,0x68,0x38,0x1C,0x16,0x0C,0x0D,0x06,0x02,0x03,0x00,0x00,0x00,0x00,
-    0x0E,0x1F,0x18,0x30,0x31,0x33,0xF2,0x7A,0x19,0x1C,0x0C,0x8C,0x86,0x02,0x21,0x01,
-    0x00,0x90,0x19,0x07,0x06,0x9C,0xF8,0x18,0x18,0x1C,0x1E,0x33,0xE1,0x81,0x01,0x03,
-    0x07,0x06,0x0E,0x0C,0x18,0x70,0xE0,0x80,0x80,0x81,0x83,0x83,0xC6,0x7E,0x38,0x00,
-
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0F,0x39,0x30,0x30,0x21,0x30,
-    0x30,0x30,0x30,0x20,0x30,0x1E,0x1E,0x3E,0xF8,0xF0,0xF4,0xF4,0xF4,0xF6,0xB0,0x1C,
-    0x03,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0xE2,0x5F,0x9A,0xAA,0x5F,0xC1,0xF8,0xFB,0x4E,0x23,0x8F,
-    0xC3,0x60,0x30,0x90,0xD8,0xC8,0xEC,0xA4,0xE4,0xA6,0xE4,0x8C,0x8C,0x98,0x30,0xE0,
-    0x00,0x00,0x00,0x00,0x00,0x0F,0x1C,0x30,0x31,0x13,0x19,0x0D,0xAC,0xFE,0x07,0x01,
-    0x01,0x00,0x00,0x00,0xE0,0xFE,0xED,0x80,0x00,0x00,0x00,0x00,0x00,0x07,0x7F,0xFE,
-    0x18,0x18,0x30,0x30,0x20,0x30,0x19,0x1F,0x05,0x01,0x01,0x01,0x00,0x00,0x00,0x00,
-
-    0x00,0x00,0x00,0x80,0xC0,0x60,0x60,0x20,0x30,0x20,0x20,0x30,0x20,0x30,0x20,0x20,
-    0x20,0x60,0x60,0xC0,0x00,0x00,0x00,0x00,0x00,0x03,0xFF,0xFF,0x00,0xEB,0x7F,0x00,
-    0x00,0x80,0xC0,0x40,0x60,0x20,0x90,0x18,0x08,0x08,0x88,0xCC,0x6C,0x34,0x1C,0x08,
-    0x00,0x00,0x00,0x00,0xEC,0xBF,0x01,0xF1,0xFF,0xFE,0xFF,0xBB,0x7E,0x1E,0x0E,0x05,
-    0x02,0x01,0x01,0x01,0x01,0x00,0x00,0x01,0x00,0x00,0x01,0x00,0x01,0x03,0x07,0x07,
-    0x00,0x00,0x00,0x70,0xE0,0xC0,0x80,0x80,0x00,0x00,0x00,0x00,0x00,0x03,0x07,0x04,
-    0x04,0x06,0x02,0x03,0x00,0x0F,0xFF,0x9F,0x3F,0xFE,0x0C,0x06,0x02,0x02,0x03,0x01,
-    0x00,0x00,0x00,0x00,0x00,0x80,0xC0,0xE0,0x70,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-
-    0x00,0x03,0x01,0x03,0x03,0x03,0x03,0x03,0x02,0x06,0x06,0x06,0x0C,0x18,0x18,0x70,
-    0xEC,0x9E,0x9E,0x7C,0xF9,0xB3,0xC6,0x8C,0x38,0xE0,0xFF,0x81,0xF8,0x7F,0x98,0xC6,
-    0xE3,0x78,0x3C,0x1C,0x8E,0xC3,0x61,0x18,0x0C,0x03,0x01,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x05,0x1F,0xFC,0xFF,0xCF,0x39,0x07,0x01,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x1F,0x78,0xC1,0xC1,0x01,0x0B,0x33,0x12,0x42,0x42,0x46,0xC6,
-    0xC6,0x84,0x0C,0x18,0x30,0xE0,0xD9,0xFB,0xFA,0xEF,0x60,0x30,0x10,0x18,0x18,0x8C,
-    0x0C,0xE4,0x06,0x06,0x03,0xE3,0x3F,0x07,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,        // Page 7 (bottom)
-    0x00,0x00,0x03,0x03,0x0F,0x0D,0x1B,0x1B,0x33,0x32,0x37,0x2F,0x33,0x32,0x13,0x11,
-    0x18,0x0C,0x06,0x02,0x01,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x0F,0x0F,0x07,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x01,0x03,0x07,0x06,0x0C,0x08,0x18,0x30,0x30,
-    0x60,0x60,0x61,0xE6,0xEC,0x7E,0x7F,0x47,0x9B,0xCC,0x84,0x86,0xC2,0x43,0x41,0x61,
-    0x20,0x30,0x1C,0x06,0x03,0x01,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-
-};
-
 static const uint8_t logo16_glcd_bmp[] =
 {
 	0b00000000, 0b11000000,
@@ -168,13 +91,12 @@ static void GreenLedThread(void *arg)
 
 static void OrangeLedThread(void *arg)
 {
-	I2C_HandleTypeDef *i2cHandle = I2C_returnHandlePtr();
-	SSD1306_init(i2cHandle, SSD1306_7BIT_ADDR);
+	SSD1306_init();
 
 	while(1)
 	{
 		HAL_GPIO_TogglePin(GPIOD, GPIO_PIN_13);
 		vTaskDelay(pdMS_TO_TICKS(1000));
-		SSD1306_writeDataBuffer(i2cHandle, SSD1306_7BIT_ADDR, global_display_buffer, 1024);
+		SSD1306_writeDataBuffer(SSD1306_getDisplayBufferHandle(), 1024);
 	}
 }
